Update recursion to restart the game
Did an update to the logic around the input that would ask the user if they wanted to play again after the game ended. Instead of the input function being used twice. I adopted the course instructors method of using a while loop to clear the screen and call the function to playgame.